drm: i915: add a driver parameter(ddi_crtc_mask) for setting the crtc mask of each DDI port.
Setting the crtc mask of each DDI port is needed in some case that we
want to swap the display order. for example, we want the display monitor
which is connect to PORT_B showing up as the first display in the
system, we can set ddi_crtc_mask to 0x0712,  every four bits stands for
a crtc mask of the corresponed DDI port,  port A is at the lowest four
bits.

Signed-off-by: Cai Yulong <yulongc@hwtc.com.cn>